Refactor windows to spearate module and add file section
