+++ gui/src/main_contents/__init__.py
@@ -1,0 +1,4 @@
+from . interfaces import AbstractMainContent
+from . home_page import HomePageContent
+from . clustering_page import ClusteringPageContent
+from . clustering_sections.file_picker import FileList
