+++ gui/src/main_contents/clustering_sections/__init__.py
@@ -1,0 +1,2 @@
+from . interfaces import AbstractSection
+from . file_picker import SelectFilesSection
